Render rework

+++ src/matrix/matrix.h
@@ -5,17 +5,17 @@
 
 typedef struct s_matrix4
 {
-	float	data[4][4];
+	double	data[4][4];
 }	t_matrix4;
 
 typedef struct s_matrix3
 {
-	float	data[3][3];
+	double	data[3][3];
 }	t_matrix3;
 
 typedef struct s_matrix2
 {
-	float	data[2][2];
+	double	data[2][2];
 }	t_matrix2;
 
 typedef enum e_axis
@@ -27,12 +27,12 @@ typedef enum e_axis
 
 typedef struct s_shearing
 {
-	float	xy;
-	float	xz;
-	float	yx;
-	float	yz;
-	float	zx;
-	float	zy;
+	double	xy;
+	double	xz;
+	double	yx;
+	double	yz;
+	double	zx;
+	double	zy;
 }	t_shearing;
 
 // m4default: Get the default matrix
@@ -73,36 +73,36 @@ t_matrix2	m2subm(t_matrix3 m, int row, int col);
 //  @param row The row to remove
 //  @param col The column to remove
 //  @return The minor
-float		m3minor(t_matrix3 m, int row, int col);
+double		m3minor(t_matrix3 m, int row, int col);
 //	m2det: Get the determinant of a t_matrix2
 //	@param m The matrix
 //	@return The determinant
-float		m2det(t_matrix2 m);
+double		m2det(t_matrix2 m);
 //	m3cofactor: Get the cofactor of a t_matrix3
 //	@param m The matrix
 //	@param row The row to remove
 //	@param col The column to remove
 //	@return The cofactor
-float		m3cofactor(t_matrix3 m, int row, int col);
+double		m3cofactor(t_matrix3 m, int row, int col);
 //	m3det: Get the determinant of a t_matrix3
 //	@param m The matrix
 //	@return The determinant
-float		m3det(t_matrix3 m);
+double		m3det(t_matrix3 m);
 //	m4minor: Get the minor of a t_matrix4
 //	@param m The matrix
 //	@param row The row to remove
 //	@param col The column to remove
 //	@return The minor
-float		m4minor(t_matrix4 m, int row, int col);
+double		m4minor(t_matrix4 m, int row, int col);
 //	m4cofactor: Get the cofactor of a t_matrix4
 //	@param m The matrix
 //	@param row The row to remove
 //	@param col The column to remove
-float		m4cofactor(t_matrix4 m, int row, int col);
+double		m4cofactor(t_matrix4 m, int row, int col);
 //	m4det: Get the determinant of a t_matrix4
 //	@param m The matrix
 //	@return The determinant
-float		m4det(t_matrix4 m);
+double		m4det(t_matrix4 m);
 //	m4invert: Get the inverse of a t_matrix4
 //	@param m The matrix
 //	@param status The status of the inversion
@@ -120,7 +120,7 @@ t_matrix4	m4scaling(t_tuple4 t);
 //	@param radian The angle in radian
 //	@param axis The axis to rotate
 //	@return The rotation matrix
-t_matrix4	m4rotating(float radian, t_axis axis);
+t_matrix4	m4rotating(double radian, t_axis axis);
 //	m4shearing: Get the shearing matrix of a t_shearing
 //	@param sh The shearing values
 //	@return The shearing matrix
