+++ src/objects/sphere.h
@@ -7,7 +7,7 @@
 typedef struct s_sphere
 {
 	t_point3	origin;
-	float		radius;
+	double		radius;
 }	t_sphere;
 
 //  new_sphere: Create a new sphere object
@@ -15,6 +15,6 @@ typedef struct s_sphere
 //  @param radius The radius of the sphere
 //	@param color The color of the sphere
 //  @return A new sphere object
-t_object		*new_sphere(t_point3 origin, float radius, t_color color);
+t_object		*new_sphere(t_point3 origin, double radius, t_color color);
 
 #endif
