+++ src/vectors/vectors_properties.c
@@ -1,6 +1,6 @@
 #include "vectors.h"
 
-float	vlength2(t_vector3 v)
+double	vlength2(t_vector3 v)
 {
 	if (DEBUG && v.w == POINT)
 		printf("Warning: Getting the squared length of a point\n");
@@ -9,11 +9,11 @@ float	vlength2(t_vector3 v)
 
 t_vector3	vnormalized(t_vector3 v)
 {
-	float	invlen;
+	double	invlen;
 
 	if (DEBUG && v.w == POINT)
 		printf("Warning: Normalizing a point\n");
-	invlen = 1 / sqrtf(vlength2(v));
+	invlen = 1 / sqrt(vlength2(v));
 	return ((t_vector3){
 		v.x * invlen,
 		v.y * invlen,
@@ -24,11 +24,11 @@ t_vector3	vnormalized(t_vector3 v)
 
 t_vector3	*vnormalize(t_vector3 *v)
 {
-	float	invlen;
+	double	invlen;
 
 	if (DEBUG && v->w == POINT)
 		printf("Warning: Normalizing a point\n");
-	invlen = 1 / sqrtf(vlength2(*v));
+	invlen = 1 / sqrt(vlength2(*v));
 	v->x *= invlen;
 	v->y *= invlen;
 	v->z *= invlen;
